fix: move topbar to shared components
